Obstacle Avoidance seems proper
Each robot has 4 obstacles around they need to avoid or they terminate. Also added a reward to encourage keeping their space. Created a custom-lidar based termination, but we shouldn't need it.

Now, we're looking in to how to better visualize the obstacles and how to randomize them more (then how to add more)

Co-Authored-By: Adam Gans <130410191+AdamTheGans@users.noreply.github.com>